refactor: extract post validation in reactions to concern
